[Core] Refactor timings bucketing into own class
